#217: fix Bank Transfer Advanced preauthorization requests
