+++ service/src/main/java/com/commercetools/pspadapter/payone/transaction/paymentinadvance/BankTransferInAdvanceAuthorizationTransactionExecutor.java
@@ -1,0 +1,29 @@
+package com.commercetools.pspadapter.payone.transaction.paymentinadvance;
+
+import com.commercetools.pspadapter.payone.domain.ctp.PaymentWithCartLike;
+import com.commercetools.pspadapter.payone.domain.payone.PayonePostService;
+import com.commercetools.pspadapter.payone.domain.payone.model.common.AuthorizationRequest;
+import com.commercetools.pspadapter.payone.mapping.PayoneRequestFactory;
+import com.google.common.cache.LoadingCache;
+import io.sphere.sdk.client.BlockingSphereClient;
+import io.sphere.sdk.types.Type;
+
+import javax.annotation.Nonnull;
+
+import static io.sphere.sdk.payments.TransactionType.AUTHORIZATION;
+
+public class BankTransferInAdvanceAuthorizationTransactionExecutor extends BaseBankTransferInAdvanceTransactionExecutor {
+
+    public BankTransferInAdvanceAuthorizationTransactionExecutor(@Nonnull LoadingCache<String, Type> typeCache,
+                                                                 @Nonnull PayoneRequestFactory requestFactory,
+                                                                 @Nonnull PayonePostService payonePostService,
+                                                                 @Nonnull BlockingSphereClient client) {
+        super(AUTHORIZATION, typeCache, requestFactory, payonePostService, client);
+    }
+
+    @Nonnull
+    @Override
+    protected AuthorizationRequest createRequest(@Nonnull PaymentWithCartLike paymentWithCartLike) {
+        return requestFactory.createPreauthorizationRequest(paymentWithCartLike);
+    }
+}
